+++ Directory.Packages.props
@@ -1,34 +1,31 @@
 <Project>
-  <PropertyGroup>
-    <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageVersion Include="Azure.Identity" Version="1.14.0" />
-    <PackageVersion Include="CsvHelper" Version="33.0.1" />
-    <PackageVersion Include="ExcelMapper" Version="6.0.609" />
-    <PackageVersion Include="FakeItEasy" Version="8.3.0" />
-    <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.5" />
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.5" />
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.5" />
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.5" />
-    <PackageVersion Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
-    <PackageVersion Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="7.0.0" />
-    <PackageVersion Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
-    <PackageVersion Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="7.0.5" />
-    <PackageVersion Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
-    <PackageVersion Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageVersion Include="Grpc.AspNetCore" Version="2.71.0" />
-    <PackageVersion Include="Shouldly" Version="4.3.0" />
-    <PackageVersion Include="Testcontainers.MsSql" Version="4.4.0" />
-    <PackageVersion Include="xunit.v3" Version="2.0.2" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageVersion Include="FluentAssertions" Version="8.2.0" />
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
-    <PackageVersion Include="xunit" Version="2.9.3" />
-    <PackageVersion Include="xunit.runner.visualstudio" Version="3.1.0" />
-    <PackageVersion Include="coverlet.collector" Version="6.0.4" />
-    <PackageVersion Include="AutoBogus.Conventions" Version="2.13.1" />
-    <PackageVersion Include="AutoBogus.Moq" Version="2.13.1" />
-  </ItemGroup>
+	<PropertyGroup>
+		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageVersion Include="Azure.Identity" Version="1.14.0" />
+		<PackageVersion Include="CsvHelper" Version="33.0.1" />
+		<PackageVersion Include="ExcelMapper" Version="6.0.609" />
+		<PackageVersion Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.5" />
+		<PackageVersion Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.5" />
+		<PackageVersion Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.5" />
+		<PackageVersion Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
+		<PackageVersion Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="7.0.0" />
+		<PackageVersion Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
+		<PackageVersion Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="7.0.5" />
+		<PackageVersion Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
+		<PackageVersion Include="Grpc.AspNetCore" Version="2.71.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
+		<PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.5" />
+		<PackageVersion Include="xunit.v3" Version="2.0.2" />
+		<PackageVersion Include="xunit.runner.visualstudio" Version="3.1.0" />
+		<PackageVersion Include="coverlet.collector" Version="6.0.4" />
+		<PackageVersion Include="AutoBogus.Conventions" Version="2.13.1" />
+		<PackageVersion Include="Shouldly" Version="4.3.0" />
+		<PackageVersion Include="Testcontainers.MsSql" Version="4.4.0" />
+		<PackageVersion Include="Testcontainers.XunitV3" Version="4.4.0" />
+		<PackageVersion Include="FakeItEasy" Version="8.3.0" />
+	</ItemGroup>
 </Project>
