+++ src/Services/PeopleService/DNP.PeopleService.Tests.xUnitV3/GlobalUsings.cs
@@ -1,0 +1,7 @@
+﻿global using Xunit;
+global using Shouldly;
+global using Bogus;
+
+global using System.Text.Json;
+global using System.Net;
+global using Microsoft.EntityFrameworkCore;
