+++ src/Services/PeopleService/DNP.PeopleService.Tests.xUnitV3/ServiceTestCollection.cs
@@ -1,0 +1,7 @@
+﻿namespace DNP.PeopleService.Tests.xUnitV3;
+
+[CollectionDefinition(nameof(ServiceTestCollection))]
+public sealed class ServiceTestCollection : ICollectionFixture<ServiceTestCollectionFixture>
+{
+    
+}
