+++ src/Services/PeopleService/DNP.PeopleService.Tests.xUnitV3/xunit.runner.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json",
+  "parallelizeAssembly": false,
+  "parallelizeTestCollections": false,
+  "showLiveOutput": false,
+  "stopOnFail": false,
+  "methodDisplay": "classAndMethod"
+}
