--- src/Services/PeopleService/DNP.PeopleService.Tests/DNP.PeopleService.Tests.csproj
@@ -1,45 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<IsPackable>false</IsPackable>
-		<IsTestProject>true</IsTestProject>
-		
-		<UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
-		<TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
-		<TestingPlatformShowTestsFailure>true</TestingPlatformShowTestsFailure>
-		<TestingPlatformCaptureOutput>false</TestingPlatformCaptureOutput>
-		<EnableMSTestRunner>true</EnableMSTestRunner>
-
-	</PropertyGroup>
-
-	<ItemGroup>
-		<Content Include="xunit.runner.json" CopyToOutputDirectory="PreserveNewest" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Using Include="Xunit" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="FakeItEasy" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" />
-		<PackageReference Include="AutoBogus.Conventions" />
-		<PackageReference Include="Shouldly" />
-		<PackageReference Include="Testcontainers.MsSql" />
-		<PackageReference Include="xunit.runner.visualstudio">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-		<PackageReference Include="coverlet.collector">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-		<PackageReference Include="xunit.v3" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\DNP.PeopleService\DNP.PeopleService.csproj" />
-	</ItemGroup>
-
-</Project>
