--- src/Services/PeopleService/DNP.PeopleService.Tests/GlobalUsings.cs
@@ -1,2 +1,0 @@
-﻿global using Xunit;
-global using Shouldly;
