--- src/Services/PeopleService/DNP.PeopleService.Tests/IntegrationTestBase.cs
@@ -1,5 +1,0 @@
-﻿
-namespace DNP.PeopleService.Tests;
-public class IntegrationTestBase(PersonalServiceTestCollectionFixture testCollectionFixture, ITestOutputHelper testOutput) : PeopleServiceTestBase(testCollectionFixture, testOutput)
-{
-}
