--- src/Services/PeopleService/DNP.PeopleService.Tests/xunit.runner.json
@@ -1,3 +1,0 @@
-{
-
-}
